+++ Microsoft.Azure.Cosmos/src/Query/v3Query/QueryIterator.cs
@@ -5,6 +5,7 @@
 namespace Microsoft.Azure.Cosmos.Query
 {
     using System;
+    using System.Collections.Generic;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Cosmos.CosmosElements;
@@ -21,23 +22,28 @@ namespace Microsoft.Azure.Cosmos.Query
         private readonly CosmosQueryExecutionContext cosmosQueryExecutionContext;
         private readonly CosmosSerializationFormatOptions cosmosSerializationFormatOptions;
         private readonly RequestOptions requestOptions;
+        private readonly CosmosClientContext clientContext;
 
         private QueryIterator(
             CosmosQueryContextCore cosmosQueryContext,
             CosmosQueryExecutionContext cosmosQueryExecutionContext,
             CosmosSerializationFormatOptions cosmosSerializationFormatOptions,
-            RequestOptions requestOptions)
+            RequestOptions requestOptions,
+            CosmosClientContext clientContext)
         {
             this.cosmosQueryContext = cosmosQueryContext ?? throw new ArgumentNullException(nameof(cosmosQueryContext));
             this.cosmosQueryExecutionContext = cosmosQueryExecutionContext ?? throw new ArgumentNullException(nameof(cosmosQueryExecutionContext));
             this.cosmosSerializationFormatOptions = cosmosSerializationFormatOptions;
             this.requestOptions = requestOptions;
+            this.clientContext = clientContext ?? throw new ArgumentNullException(nameof(clientContext));
         }
 
         public static QueryIterator Create(
             CosmosQueryClient client,
+            CosmosClientContext clientContext,
             SqlQuerySpec sqlQuerySpec,
             string continuationToken,
+            FeedRangeInternal feedRangeInternal,
             QueryRequestOptions queryRequestOptions,
             Uri resourceLink,
             bool isContinuationExpected,
@@ -75,9 +81,10 @@ namespace Microsoft.Azure.Cosmos.Query
                                 cosmosQueryContext,
                                 new QueryExecutionContextWithException(
                                     new MalformedContinuationTokenException(
-                                        $"Malformed Continuation Token: {requestContinuationToken}")),
+                                        $"Malformed Continuation Token: {continuationToken}")),
                                 queryRequestOptions.CosmosSerializationFormatOptions,
-                                queryRequestOptions);
+                                queryRequestOptions,
+                                clientContext);
                         }
                     }
                     else
@@ -97,6 +104,7 @@ namespace Microsoft.Azure.Cosmos.Query
             CosmosQueryExecutionContextFactory.InputParameters inputParameters = new CosmosQueryExecutionContextFactory.InputParameters(
                 sqlQuerySpec: sqlQuerySpec,
                 initialUserContinuationToken: requestContinuationToken,
+                initialFeedRange: feedRangeInternal,
                 maxConcurrency: queryRequestOptions.MaxConcurrency,
                 maxItemCount: queryRequestOptions.MaxItemCount,
                 maxBufferedItemCount: queryRequestOptions.MaxBufferedItemCount,
@@ -111,17 +119,11 @@ namespace Microsoft.Azure.Cosmos.Query
                 cosmosQueryContext,
                 CosmosQueryExecutionContextFactory.Create(cosmosQueryContext, inputParameters),
                 queryRequestOptions.CosmosSerializationFormatOptions,
-                queryRequestOptions);
+                queryRequestOptions,
+                clientContext);
         }
 
         public override bool HasMoreResults => !this.cosmosQueryExecutionContext.IsDone;
-
-#if PREVIEW
-        public override
-#else
-        internal
-#endif
-        FeedToken FeedToken => throw new NotImplementedException();
 
         public override async Task<ResponseMessage> ReadNextAsync(CancellationToken cancellationToken = default)
         {
@@ -136,8 +138,14 @@ namespace Microsoft.Azure.Cosmos.Query
 
                 if (responseCore.IsSuccess)
                 {
+                    List<CosmosElement> decryptedCosmosElements = null;
+                    if (this.clientContext.ClientOptions.Encryptor != null)
+                    {
+                        decryptedCosmosElements = await this.GetDecryptedElementResponseAsync(responseCore.CosmosElements, diagnostics, cancellationToken);
+                    }
+
                     return QueryResponse.CreateSuccess(
-                        result: responseCore.CosmosElements,
+                        result: decryptedCosmosElements ?? responseCore.CosmosElements,
                         count: responseCore.CosmosElements.Count,
                         responseLengthBytes: responseCore.ResponseLengthBytes,
                         diagnostics: diagnostics,
@@ -167,8 +175,8 @@ namespace Microsoft.Azure.Cosmos.Query
                     responseHeaders: new CosmosQueryResponseMessageHeaders(
                         responseCore.ContinuationToken,
                         responseCore.DisallowContinuationTokenMessage,
-                        cosmosQueryContext.ResourceTypeEnum,
-                        cosmosQueryContext.ContainerResourceId)
+                        this.cosmosQueryContext.ResourceTypeEnum,
+                        this.cosmosQueryContext.ContainerResourceId)
                     {
                         RequestCharge = responseCore.RequestCharge,
                         ActivityId = responseCore.ActivityId,
@@ -180,6 +188,35 @@ namespace Microsoft.Azure.Cosmos.Query
         public override CosmosElement GetCosmsoElementContinuationToken()
         {
             return this.cosmosQueryExecutionContext.GetCosmosElementContinuationToken();
+        }
+
+        private async Task<List<CosmosElement>> GetDecryptedElementResponseAsync(
+            IReadOnlyList<CosmosElement> encryptedCosmosElements,
+            CosmosDiagnosticsContext diagnosticsContext,
+            CancellationToken cancellationToken)
+        {
+            List<CosmosElement> decryptedCosmosElements = new List<CosmosElement>();
+            using (diagnosticsContext.CreateScope("Decrypt"))
+            {
+                foreach (CosmosElement document in encryptedCosmosElements)
+                {
+                    if (!(document is CosmosObject documentObject))
+                    {
+                        decryptedCosmosElements.Add(document);
+                        continue;
+                    }
+
+                    CosmosObject decryptedDocument = await this.clientContext.EncryptionProcessor.DecryptAsync(
+                        documentObject,
+                        this.clientContext.ClientOptions.Encryptor,
+                        diagnosticsContext,
+                        cancellationToken);
+
+                    decryptedCosmosElements.Add(decryptedDocument);
+                }
+            }
+
+            return decryptedCosmosElements;
         }
     }
 }