+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Json/JsonWriterTests.cs
@@ -1746,7 +1746,7 @@
 
                         case JsonTokenType.Number:
                             Number64 numberValue = (token as JsonNumberToken).Value;
-                            jsonWriter.WriteNumberValue(numberValue);
+                            jsonWriter.WriteNumber64Value(numberValue);
                             break;
 
                         case JsonTokenType.True: